requirements.txt
requests==2.31.0
pandas==2.2.2
google-cloud-storage==2.16.0
python-dotenv==1.0.1
pyspark==3.5.1